dom/camera/test/callback/test_bug1099390.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1099390</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id="viewfinder" width="200" height="200" autoplay></video>
<img src="#" alt="This image is going to load" id="testimage"/>
<script class="testbody" type="text/javascript;version=1.7">

var whichCamera = navigator.mozCameras.getListOfCameras()[0];
var config = {
  mode: 'picture',
  recorderProfile: 'cif',
  previewSize: {
    width: 352,
    height: 288
  }
};

function onError(e) {
  ok(false, "Error " + e);
}

var Camera = {
  _cameraObj: null,

  get viewfinder() {
    return document.getElementById('viewfinder');
  },

  release: function release() {
    viewfinder.mozSrcObject = null;
    if (Camera._cameraObj) {
      Camera._cameraObj.release();
      Camera._cameraObj = null;
    }
  },

  test: function test(cam) {
    var gotCloseEvent = false;
    var gotReleaseCallback = false;

    function gotAll() {
      var all = gotCloseEvent && gotReleaseCallback;
      if (all) {
        info("Got all expected notifications");
      }
      return all;
    };

    cam.onClosed = function(reason) {
      cam.onClosed = null;
      ok(!gotCloseEvent, "gotCloseEvent was " + gotCloseEvent);
      ok(reason === "HardwareReleased", "onClosed reason is: " + reason);
      gotCloseEvent = true;
      if (gotAll()) {
        SimpleTest.finish();
      }
    };

    var onSuccess = function() {
      ok(!gotReleaseCallback, "gotReleaseCallback was " + gotReleaseCallback);
      gotReleaseCallback = true;
      if (gotAll()) {
        SimpleTest.finish();
      }
    };

    cam.release(onSuccess, onError);
  }, // test()

  start: function start() {
    function onSuccess(cam) {
      Camera._cameraObj = cam;
      Camera.viewfinder.mozSrcObject = cam;
      Camera.viewfinder.play();

      var onPreviewStateChange = function(e) {
        if (e.newState === 'started') {
          cam.removeEventListener('previewstatechange', onPreviewStateChange);
          Camera.test(cam);
        }
      }; // onPreviewStateChange
      cam.addEventListener('previewstatechange', onPreviewStateChange);
    }; // onSuccess()

    navigator.mozCameras.getCamera(whichCamera, config, onSuccess, onError);
  }, // start()
}

SimpleTest.waitForExplicitFinish();

window.addEventListener('beforeunload', function() {
  Camera.release();
});

Camera.start();

</script>
</body>

</html>
